requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
joblib
locust
torchserve 
torch-model-archiver